University-Time-table-scheduler-main/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mongoengine==0.27.0
pymongo==4.6.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
reportlab==4.0.7
openpyxl==3.1.2
Pillow==11.0.0
django-filter==23.3
django-rest-framework-mongoengine==3.4.1
setuptools<81
tqdm